requirements.txt
aniso8601==3.0.0
click==6.7
flask-restplus==0.10.1
flask==1.0.2
itsdangerous==0.24
jinja2==2.10
jsonschema==2.6.0
markupsafe==1.0
pytz==2018.4
six==1.11.0
werkzeug==0.14.1
pytest==3.5.1
pytest-cov==2.5.1
pyyaml
opcua
numpy
pyaml
pyopenssl
